Move `Bencher` to a separate `bencher` module
